[bug fixing][important note]
In this commit, I enhanced the reordering module. The problem of reorder
was caused by the ambiguous view creation logic. More specifically, the
problem is when should I create view models and what is the order of
view creation. First, an EnforceNodeStatement may have an abstract
target type. However, such an EnforceNodeStatement is expected to be
refined by another one that has a non-abstract target type. This case
can be handled by postponing non-enforceable statements. Second, a
RuleCallStatement may either return view elements or require view
elements. The manner cannot be determined statically at present. Third,
an inner ForEachStatement may require that a view element (i.e., the
root element) should be created before matching the view pattern. For
the first case, I have to collect all the possible RuleCallStatemets and
call them in the trial mode before creating any view element in order to
avoid duplicate creation; for the second and the third case, I have to
create view elements before calling rules and executing
ForEachStatements--two entire different semantics. 

The reordering problem becomes more complicated when CaseStatements and
ForEachStatements are nested.

To handle this problem, the current reordering logic is to create a
flatten structure. Every the possible execution path is collected and
restructured as a tuple of (precondition, match, enforce, lazy, tail).
In principle, precondition is the condition collected from
CaseStatements; match includes the all MatchPattern statements; enforce
includes executable view creation statements; lazy includes the the view
creation statements and foreach statements that should be postponed
(recursive dependencies are not supported); tail includes the rule
calls. Currently, I cannot assure the correctness of this reordering
logic.

Another possible solution is to allow dynamic up-casting, i.e., we can
permit that a view element can be dynamically casted into a subtype when
it is just created (in the same rule). However, I also don't know the
correctness of this solution.

The ultimate solution is that we must be able to statically determine
the execution order.